Переделал все компоненты в классы
